+++ lib/preproc.go
@@ -26,10 +26,14 @@ func Preproc(str string) string {
 	for i := range lines {
 		indent := indentCount(lines[i])
 		token := "=>"
+		eqIf := "= if "
 		if len(lines[i]) > indent+4 && lines[i][indent:indent+4] == "for " {
 			lines[i] += ";"
 		}
 		if len(lines[i]) > indent+3 && lines[i][indent:indent+3] == "if " && !strings.Contains(lines[i], token) {
+			lines[i] += ";"
+		}
+		if len(lines[i]) > indent+3 && strings.Contains(lines[i], eqIf) && !strings.Contains(lines[i], token) {
 			lines[i] += ";"
 		}
 		if len(lines[i]) > indent+5 && lines[i][indent:indent+5] == "else " && !strings.Contains(lines[i], token) {